use narrowest type

+++ firmware/patch.c
@@ -422,7 +422,7 @@ void start_dsp_thread(void) {
 
 
 void computebufI(int32_t *inp, int32_t *outp) {
-    int i; for (i = 0; i < 32; i++) {
+    uint8_t i; for (i = 0; i < 32; i++) {
         inbuf[i] = inp[i];
     }
